refactor: restructure bridge call logic for improved readability
